printf: Start move to being in GPU-AV
