+++ layers/gpu/core/gpu_state_tracker.h
@@ -16,33 +16,14 @@
  */
 #pragma once
 #include "state_tracker/cmd_buffer_state.h"
-#include "state_tracker/queue_state.h"
 
 namespace gpu {
 class GpuShaderInstrumentor;
 }
 
+// TODO - Remove this file and namespace
+// The only reason we have this is because DebugPrintf and GPU-AV have different command buffers
 namespace gpu_tracker {
-
-class Queue : public vvl::Queue {
-  public:
-    Queue(gpu::GpuShaderInstrumentor &shader_instrumentor_, VkQueue q, uint32_t family_index, uint32_t queue_index,
-          VkDeviceQueueCreateFlags flags, const VkQueueFamilyProperties &queueFamilyProperties, bool timeline_khr);
-    virtual ~Queue();
-
-  protected:
-    vvl::PreSubmitResult PreSubmit(std::vector<vvl::QueueSubmission> &&submissions) override;
-    void PostSubmit(vvl::QueueSubmission &) override;
-    void SubmitBarrier(const Location &loc, uint64_t seq);
-    void Retire(vvl::QueueSubmission &) override;
-
-    gpu::GpuShaderInstrumentor &shader_instrumentor_;
-    VkCommandPool barrier_command_pool_{VK_NULL_HANDLE};
-    VkCommandBuffer barrier_command_buffer_{VK_NULL_HANDLE};
-    VkSemaphore barrier_sem_{VK_NULL_HANDLE};
-    std::deque<std::vector<std::shared_ptr<vvl::CommandBuffer>>> retiring_;
-    const bool timeline_khr_;
-};
 
 class CommandBuffer : public vvl::CommandBuffer {
   public:
@@ -54,5 +35,4 @@ class CommandBuffer : public vvl::CommandBuffer {
 };
 }  // namespace gpu_tracker
 
-VALSTATETRACK_DERIVED_STATE_OBJECT(VkQueue, gpu_tracker::Queue, vvl::Queue)
 VALSTATETRACK_DERIVED_STATE_OBJECT(VkCommandBuffer, gpu_tracker::CommandBuffer, vvl::CommandBuffer)
